PlasmodeCont.R sampling ID error
In functions where exposure/outcome model formulas are provided, [X] and [Xbnew] are estimated using the original user-provided [data]. However, sampling of each sim in the for loop is chosen based on the row number of [x] which is [data] sorted by exposure value. This creates a mismatch when assigning the simulated outcome [ynew], by pulling the wrong value from [Xbnew]. 

A symptom is that all simulation sets will usually converge to a RD ~0, regardless of the effectOR specified or input models. I have confirmed these edits work for the versions where formulas are provided. 

It is possible that functions which take an input model may also be affected, if there is a mismatch between the matrix the object is built from and how it is utilized in the function. I have pointed these lines out, but have not had time to consider if it needs to be addressed  and how.